+++ src/tests/streams/writable/main.js
@@ -1,0 +1,64 @@
+// Copyright (c) 2026 Cloudflare, Inc.
+// Licensed under the Apache 2.0 license found in the LICENSE file or at:
+//     https://opensource.org/licenses/Apache-2.0
+
+// Entry point for the writable streams suite. Re-exports are explicit
+// (not `export *`) so a name collision between modules is a load-time
+// SyntaxError instead of a silently dropped test.
+
+export {
+  writableGlobalsExist,
+  controllerNotConstructable,
+  newWritableStream,
+} from 'api-surface';
+
+export { highWaterMarkValidated } from 'construction';
+
+export {
+  newWritableStreamWithSink,
+  newWritableStreamWithSinkAsync,
+  newWritableStreamStartError,
+  newWritableStreamWriteError,
+  newWritableStreamAbortError,
+  newWritableStreamCloseError,
+  writableStreamSizeAlgorithm,
+} from 'sink-algorithms';
+
+export {
+  writableStreamMultiplePendingWrites,
+  writableStreamWriteSubarray,
+  writableStreamWriteAny,
+  writableStreamPromisesResolvedInOrder,
+} from 'write-semantics';
+
+export {
+  writableStreamCloseThrowRejectsPromises,
+  writerDoubleClose,
+} from 'close-semantics';
+
+export {
+  writableStreamWriteAbort,
+  writableStreamAbortReadyRejected,
+  writableStreamAbortOptional,
+  writableStreamAbortWhileStarting,
+  writableStreamAbortWhileWriting,
+  writableStreamReleaseLockWhileAborting,
+  writableStreamAbortTiming,
+  writableStreamAbortWriteClosePending,
+  writableStreamErrorDuringInFlightWrite,
+  writableStreamStartErrorAfterAbort,
+  writableStreamRejectedWriteNoPreventAbort,
+  writableStreamAbortedTwice,
+  writableStreamAbortOnErroredResolves,
+  writableStreamSinkAlgNoCallErrorBeforeAbort,
+  writableStreamWriterWithPendingAbort,
+  errorRaceWithCloseWritable,
+} from 'abort-semantics';
+
+export {
+  writableStreamDesiredSize,
+  backpressureWritableDesiredSize,
+  backpressureWritableSlowSink,
+} from 'backpressure';
+
+export { writableStreamGc } from 'gc';
