+++ src/tests/streams/writable/writable-ts.wd-test
@@ -1,0 +1,28 @@
+using Workerd = import "/workerd/workerd.capnp";
+using WritableModules = import "writable-modules.capnp";
+
+# Runs the writable streams suite against the TypeScript implementation,
+# with byte-for-byte the same modules as writable-cpp.wd-test. The flags
+# pinned there are deliberately absent: the TypeScript implementation
+# hard-codes the modern behaviors, and the `@` / `@all-compat-flags`
+# variants verify that indifference with the flags all off and all on.
+
+const unitTests :Workerd.Config = (
+  services = [
+    ( name = "streams-writable-ts",
+      worker = (
+        modules = WritableModules.modules,
+        compatibilityFlags = [
+          "nodejs_compat",
+          "typescript_implemented_streams",
+          "experimental",
+        ],
+      )
+    ),
+  ],
+  autogates = [
+    "workerd-autogate-per-isolate-javascript-bootstrap",
+  ],
+  # For gc.js.
+  v8Flags = ["--expose-gc"],
+);
