+++ src/workerd/api/tests/streams-backpressure-test.js
@@ -135,108 +135,6 @@ export const backpressureReadableHwmLarge = {
   },
 };
 
-// Test WritableStream desiredSize tracking and ready promise
-// Inspired by: Bun test/js/web/fetch/body-stream.test.ts (backpressure with various input lengths)
-export const backpressureWritableDesiredSize = {
-  async test() {
-    const written = [];
-
-    const ws = new WritableStream(
-      {
-        write(chunk) {
-          written.push(chunk);
-          // Simulate slow write
-          return new Promise((resolve) => setTimeout(resolve, 10));
-        },
-      },
-      { highWaterMark: 3 }
-    );
-
-    const writer = ws.getWriter();
-
-    strictEqual(writer.desiredSize, 3, 'initial desiredSize');
-
-    writer.write(1);
-    strictEqual(writer.desiredSize, 2, 'desiredSize after first write');
-
-    writer.write(2);
-    strictEqual(writer.desiredSize, 1, 'desiredSize after second write');
-
-    writer.write(3);
-    strictEqual(writer.desiredSize, 0, 'desiredSize at capacity');
-
-    const readyBefore = writer.ready;
-    ok(readyBefore instanceof Promise, 'ready should be a Promise');
-
-    writer.write(4);
-    strictEqual(writer.desiredSize, -1, 'desiredSize over capacity');
-
-    const readyAfter = writer.ready;
-    ok(readyAfter instanceof Promise, 'ready should still be a Promise');
-
-    await scheduler.wait(50);
-
-    ok(writer.desiredSize > -1, 'desiredSize recovered after writes complete');
-
-    const readyResolved = writer.ready;
-    await readyResolved;
-
-    await writer.close();
-    strictEqual(written.length, 4, 'all chunks written');
-  },
-};
-
-// Test WritableStream with slow sink causes backpressure
-// Inspired by: Bun test/js/bun/spawn/spawn-stdin-readable-stream-edge-cases.test.ts (slow consumer)
-export const backpressureWritableSlowSink = {
-  async test() {
-    let writeCount = 0;
-
-    const ws = new WritableStream(
-      {
-        async write() {
-          writeCount++;
-          // Very slow write
-          await scheduler.wait(50);
-        },
-      },
-      { highWaterMark: 2 }
-    );
-
-    const writer = ws.getWriter();
-
-    strictEqual(writer.desiredSize, 2);
-
-    const w1 = writer.write('a');
-    strictEqual(writer.desiredSize, 1);
-
-    const w2 = writer.write('b');
-    strictEqual(writer.desiredSize, 0);
-
-    const initialReady = writer.ready;
-
-    const w3 = writer.write('c');
-    strictEqual(writer.desiredSize, -1);
-
-    // Backpressure can be signaled in two ways depending on implementation:
-    // 1. writer.ready returns a new pending promise (ready !== initialReady)
-    // 2. desiredSize drops to 0 or below, indicating the queue is full
-    // We accept either signal as valid backpressure indication.
-    ok(
-      writer.ready !== initialReady || writer.desiredSize <= 0,
-      'backpressure signal'
-    );
-
-    const results = await Promise.allSettled([w1, w2, w3]);
-    strictEqual(results[0].status, 'fulfilled', 'first write should succeed');
-    strictEqual(results[1].status, 'fulfilled', 'second write should succeed');
-    strictEqual(results[2].status, 'fulfilled', 'third write should succeed');
-    strictEqual(writeCount, 3);
-
-    await writer.close();
-  },
-};
-
 // Test TransformStream with both readable and writable strategies
 // Inspired by: Bun test/js/web/streams/streams.test.js (TransformStream tests)
 export const backpressureTransformBothStrategies = {
