+++ src/workerd/api/tests/streams-js-test.js
@@ -2,8 +2,9 @@
 // Licensed under the Apache 2.0 license found in the LICENSE file or at:
 //     https://opensource.org/licenses/Apache-2.0
 
-// Tests for JavaScript-backed streams (ReadableStream and WritableStream constructors)
-// Ported from edgeworker streams-js.ew-test
+// Tests for JavaScript-backed streams (ReadableStream constructors).
+// Ported from edgeworker streams-js.ew-test. The WritableStream tests
+// live in the writable suite (src/tests/streams/writable/).
 
 import { strictEqual, ok, throws, rejects } from 'node:assert';
 
@@ -13,7 +14,6 @@ export const userStreamsGlobalsExist = {
     ok(ReadableStreamDefaultController !== undefined);
     ok(ReadableByteStreamController !== undefined);
     ok(ReadableStreamBYOBRequest !== undefined);
-    ok(WritableStreamDefaultController !== undefined);
   },
 };
 
@@ -23,7 +23,6 @@ export const jsStreamsObjectsNotConstructable = {
     throws(() => new ReadableStreamDefaultController(), TypeError);
     throws(() => new ReadableByteStreamController(), TypeError);
     throws(() => new ReadableStreamBYOBRequest(), TypeError);
-    throws(() => new WritableStreamDefaultController(), TypeError);
   },
 };
 
@@ -1396,981 +1395,6 @@ export const readableStreamDefaultTee = {
 };
 
 // =====================================================================================
-// WritableStream tests
-// =====================================================================================
-
-// Test new WritableStream() works
-export const newWritableStream = {
-  test() {
-    new WritableStream();
-  },
-};
-
-// Test new WritableStream() with sink works
-export const newWritableStreamWithSink = {
-  async test() {
-    // Sync sink with abort
-    {
-      let started = false;
-      let written = false;
-      let closed = false;
-      let aborted = false;
-      const ws = new WritableStream({
-        start(c) {
-          ok(c instanceof WritableStreamDefaultController);
-          started = true;
-        },
-        write(value, c) {
-          strictEqual(value, 1);
-          ok(c instanceof WritableStreamDefaultController);
-          written = true;
-        },
-        abort(reason) {
-          strictEqual(reason.message, 'boom');
-          aborted = true;
-        },
-        close() {
-          closed = true;
-        },
-      });
-
-      ok(started);
-
-      const writer = ws.getWriter();
-
-      await writer.write(1);
-      ok(written);
-
-      await writer.abort(new Error('boom'));
-      ok(aborted);
-      ok(!closed);
-
-      await rejects(writer.closed);
-    }
-
-    // Sync sink with close
-    {
-      let started = false;
-      let written = false;
-      let closed = false;
-      let aborted = false;
-      const ws = new WritableStream({
-        start(c) {
-          ok(c instanceof WritableStreamDefaultController);
-          started = true;
-        },
-        write(value, c) {
-          strictEqual(value, 1);
-          ok(c instanceof WritableStreamDefaultController);
-          written = true;
-        },
-        abort() {
-          aborted = true;
-        },
-        close() {
-          closed = true;
-        },
-      });
-
-      ok(started);
-
-      const writer = ws.getWriter();
-
-      await writer.write(1);
-      ok(written);
-
-      await Promise.all([writer.close(), writer.closed]);
-
-      ok(!aborted);
-      ok(closed);
-    }
-  },
-};
-
-// Test new WritableStream() with async sink works
-export const newWritableStreamWithSinkAsync = {
-  async test() {
-    // Async sink with abort
-    {
-      let started = false;
-      let written = false;
-      let closed = false;
-      let aborted = false;
-      const ws = new WritableStream({
-        async start(c) {
-          ok(c instanceof WritableStreamDefaultController);
-          await scheduler.wait(10);
-          started = true;
-        },
-        async write(value, c) {
-          await scheduler.wait(10);
-          strictEqual(value, 1);
-          ok(c instanceof WritableStreamDefaultController);
-          written = true;
-        },
-        async abort(reason) {
-          await scheduler.wait(10);
-          strictEqual(reason.message, 'boom');
-          aborted = true;
-        },
-        async close() {
-          closed = true;
-        },
-      });
-
-      await scheduler.wait(15);
-      ok(started);
-
-      const writer = ws.getWriter();
-      await writer.ready;
-
-      await writer.write(1);
-      ok(written);
-
-      await writer.abort(new Error('boom'));
-      ok(aborted);
-      ok(!closed);
-
-      await rejects(writer.closed);
-    }
-
-    // Async sink with close
-    {
-      let started = false;
-      let written = false;
-      let closed = false;
-      let aborted = false;
-      const ws = new WritableStream({
-        async start(c) {
-          ok(c instanceof WritableStreamDefaultController);
-          await scheduler.wait(10);
-          started = true;
-        },
-        async write(value, c) {
-          await scheduler.wait(10);
-          strictEqual(value, 1);
-          ok(c instanceof WritableStreamDefaultController);
-          written = true;
-        },
-        async abort() {
-          await scheduler.wait(10);
-          aborted = true;
-        },
-        async close() {
-          await scheduler.wait(10);
-          closed = true;
-        },
-      });
-
-      await scheduler.wait(15);
-      ok(started);
-
-      const writer = ws.getWriter();
-      await writer.ready;
-
-      await writer.write(1);
-      ok(written);
-
-      await Promise.all([writer.close(), writer.closed]);
-      ok(!aborted);
-      ok(closed);
-    }
-  },
-};
-
-// Test new WritableStream() start algorithm error handled
-export const newWritableStreamStartError = {
-  async test() {
-    // Sync start error
-    {
-      const ws = new WritableStream({
-        start() {
-          throw new Error('boom');
-        },
-      });
-
-      const writer = ws.getWriter();
-      await rejects(writer.write(1), { message: 'boom' });
-    }
-
-    // Async start error
-    {
-      const ws = new WritableStream({
-        async start() {
-          throw new Error('boom');
-        },
-      });
-
-      const writer = ws.getWriter();
-      await rejects(writer.write(1), { message: 'boom' });
-    }
-
-    // Start with controller.error
-    {
-      const ws = new WritableStream({
-        start(c) {
-          c.error(new Error('boom'));
-        },
-      });
-
-      const writer = ws.getWriter();
-      await rejects(writer.write(1), { message: 'boom' });
-    }
-  },
-};
-
-// Test new WritableStream() write algorithm error handled
-export const newWritableStreamWriteError = {
-  async test() {
-    // Sync write error
-    {
-      const ws = new WritableStream({
-        write() {
-          throw new Error('boom');
-        },
-      });
-
-      const writer = ws.getWriter();
-      await rejects(writer.write(1), { message: 'boom' });
-    }
-
-    // Async write error
-    {
-      const ws = new WritableStream({
-        async write() {
-          throw new Error('boom');
-        },
-      });
-
-      const writer = ws.getWriter();
-      await rejects(writer.write(1), { message: 'boom' });
-    }
-
-    // Write with controller.error
-    {
-      const ws = new WritableStream({
-        write(value, c) {
-          strictEqual(value, 1);
-          c.error(new Error('boom'));
-        },
-      });
-
-      const writer = ws.getWriter();
-
-      // Should succeed
-      await writer.write(1);
-
-      await rejects(writer.closed, { message: 'boom' });
-    }
-  },
-};
-
-// Test new WritableStream() abort algorithm error handled
-export const newWritableStreamAbortError = {
-  async test() {
-    // Sync abort error
-    {
-      const ws = new WritableStream({
-        abort(reason) {
-          strictEqual(reason, 1);
-          throw new Error('boom');
-        },
-      });
-
-      const writer = ws.getWriter();
-
-      await rejects(writer.abort(1), { message: 'boom' });
-
-      // Calling abort again returns the same rejected promise
-      await rejects(writer.abort(1), { message: 'boom' });
-    }
-
-    // Async abort error
-    {
-      const ws = new WritableStream({
-        async abort(reason) {
-          strictEqual(reason, 1);
-          throw new Error('boom');
-        },
-      });
-
-      const writer = ws.getWriter();
-      await rejects(writer.abort(1), { message: 'boom' });
-    }
-
-    // Abort with controller.error
-    {
-      let controller;
-      const ws = new WritableStream({
-        start(c) {
-          controller = c;
-        },
-        async abort(reason) {
-          strictEqual(reason, 1);
-          controller.error(new Error('ignored'));
-        },
-      });
-
-      const writer = ws.getWriter();
-
-      await writer.abort(1);
-
-      // The closed promise will use the abort reason, not the error
-      // reported in the controller
-      const results = await Promise.allSettled([writer.closed]);
-      strictEqual(results[0].status, 'rejected');
-      strictEqual(results[0].reason, 1);
-    }
-  },
-};
-
-// Test new WritableStream() close algorithm error handled
-export const newWritableStreamCloseError = {
-  async test() {
-    // Sync close error
-    {
-      const ws = new WritableStream({
-        close() {
-          throw new Error('boom');
-        },
-      });
-
-      const writer = ws.getWriter();
-      await rejects(writer.close(), { message: 'boom' });
-    }
-
-    // Async close error
-    {
-      const ws = new WritableStream({
-        async close() {
-          throw new Error('boom');
-        },
-      });
-
-      const writer = ws.getWriter();
-      await rejects(writer.close(), { message: 'boom' });
-    }
-
-    // Close with controller.error (ignored)
-    {
-      let controller;
-      const ws = new WritableStream({
-        start(c) {
-          controller = c;
-        },
-        async close() {
-          controller.error(new Error('ignored'));
-        },
-      });
-
-      const writer = ws.getWriter();
-
-      // In this case, the error reported in the close algorithm is ignored
-      await Promise.all([writer.close(), writer.closed]);
-    }
-  },
-};
-
-// Test WritableStream multiple pending writes allowed
-export const writableStreamMultiplePendingWrites = {
-  async test() {
-    const expectedWrites = ['hello', 'there'];
-    const ws = new WritableStream({
-      async write(value) {
-        await scheduler.wait(10);
-        strictEqual(value, expectedWrites.shift());
-      },
-    });
-
-    const writer = ws.getWriter();
-
-    await Promise.all([writer.write('hello'), writer.write('there')]);
-  },
-};
-
-// Test WritableStream writing a subarray works
-export const writableStreamWriteSubarray = {
-  async test() {
-    const u8 = new Uint8Array([1, 2, 3, 4]);
-    const sub = u8.subarray(1, 3);
-
-    const ws = new WritableStream({
-      write(value) {
-        strictEqual(value, sub);
-      },
-    });
-
-    const writer = ws.getWriter();
-
-    await writer.write(sub);
-  },
-};
-
-// Test WritableStream writing any javascript value works
-export const writableStreamWriteAny = {
-  async test() {
-    // Make a copy since we'll shift() from it
-    const expectedWrites = [
-      'hello',
-      true,
-      1,
-      1.1,
-      undefined,
-      NaN,
-      Infinity,
-      new Uint8Array(1),
-      {},
-      [],
-    ];
-    // Keep original for writing
-    const valuesToWrite = [...expectedWrites];
-
-    const ws = new WritableStream({
-      async write(value) {
-        await scheduler.wait(1);
-        const expected = expectedWrites.shift();
-        // Use Object.is for NaN and -0/+0 handling (same as testharness same_value)
-        ok(Object.is(value, expected), `expected ${expected} but got ${value}`);
-      },
-    });
-
-    const writer = ws.getWriter();
-
-    await Promise.all(valuesToWrite.map((i) => writer.write(i)));
-  },
-};
-
-// Test WritableStream desiredSize calculated correctly
-export const writableStreamDesiredSize = {
-  async test() {
-    const ws = new WritableStream(
-      {
-        async write() {
-          await scheduler.wait(10);
-        },
-      },
-      {
-        highWaterMark: 2,
-      }
-    );
-
-    const writer = ws.getWriter();
-
-    const firstReady = writer.ready;
-    await firstReady;
-
-    strictEqual(writer.desiredSize, 2);
-    const write1 = writer.write(1);
-    strictEqual(writer.desiredSize, 1);
-
-    const write2 = writer.write(2);
-    const write3 = writer.write(3);
-
-    strictEqual(writer.desiredSize, -1);
-
-    await Promise.all([write1, write2, write3]);
-
-    ok(firstReady != writer.ready);
-    await writer.ready;
-  },
-};
-
-// Test WritableStream writes can be aborted
-export const writableStreamWriteAbort = {
-  async test() {
-    let aborted = false;
-    const ws = new WritableStream({
-      start(c) {
-        c.signal.addEventListener(
-          'abort',
-          () => {
-            strictEqual(c.signal.reason.message, 'boom');
-            aborted = true;
-          },
-          { once: true }
-        );
-      },
-      async write() {
-        await scheduler.wait(10);
-      },
-    });
-
-    const writer = ws.getWriter();
-    const write1 = writer.write(1);
-    const write2 = writer.write(2);
-
-    writer.abort(new Error('boom'));
-
-    const [result1, result2] = await Promise.allSettled([write1, write2]);
-
-    // Both writes fail
-    strictEqual(result1.status, 'rejected');
-    strictEqual(result2.status, 'rejected');
-    strictEqual(result1.reason.message, 'boom');
-    strictEqual(result2.reason.message, 'boom');
-
-    ok(aborted);
-
-    // Aborting puts the stream into a persistent errored state
-    writer.releaseLock();
-    const writer2 = ws.getWriter();
-
-    await rejects(writer2.write('should not be allowed'), { message: 'boom' });
-  },
-};
-
-// Test WritableStream uses size algorithm correctly
-export const writableStreamSizeAlgorithm = {
-  async test() {
-    // Size algorithm called
-    {
-      let sizeCalled = false;
-      const ws = new WritableStream(
-        {
-          async write() {
-            await scheduler.wait(10);
-          },
-        },
-        {
-          highWaterMark: 2,
-          size(value) {
-            sizeCalled = true;
-            strictEqual(value, 'hello');
-            return 2;
-          },
-        }
-      );
-
-      const writer = ws.getWriter();
-      strictEqual(writer.desiredSize, 2);
-      const write = writer.write('hello');
-      ok(sizeCalled);
-      strictEqual(writer.desiredSize, 0);
-      await write;
-    }
-
-    // Size algorithm error
-    {
-      const ws = new WritableStream(
-        {},
-        {
-          size() {
-            throw new Error('boom');
-          },
-        }
-      );
-
-      const writer = ws.getWriter();
-      await rejects(writer.write('hello'), { message: 'boom' });
-    }
-  },
-};
-
-// Test WritableStream aborting should replace ready promise
-export const writableStreamAbortReadyRejected = {
-  async test() {
-    const ws = new WritableStream();
-    const writer = ws.getWriter();
-    const ready = writer.ready;
-
-    await ready;
-
-    writer.abort('boom');
-
-    ok(ready !== writer.ready);
-
-    const results = await Promise.allSettled([writer.ready]);
-    strictEqual(results[0].status, 'rejected');
-    strictEqual(results[0].reason, 'boom');
-  },
-};
-
-// Test WritableStream abort with no argument defaults to undefined
-export const writableStreamAbortOptional = {
-  async test() {
-    const ws = new WritableStream();
-    const writer = ws.getWriter();
-    await writer.abort();
-
-    const results = await Promise.allSettled([writer.closed]);
-    strictEqual(results[0].status, 'rejected');
-    strictEqual(results[0].reason, undefined);
-  },
-};
-
-// Test WritableStream abort while starting rejects ready promise
-export const writableStreamAbortWhileStarting = {
-  async test() {
-    const ws = new WritableStream({
-      async start() {},
-    });
-
-    const writer = ws.getWriter();
-    writer.abort('boom');
-
-    const results = await Promise.allSettled([writer.ready]);
-    strictEqual(results[0].status, 'rejected');
-    strictEqual(results[0].reason, 'boom');
-  },
-};
-
-// Test WritableStream abort error while writing rejects abort promise
-export const writableStreamAbortWhileWriting = {
-  async test() {
-    const ws = new WritableStream({
-      async write() {
-        await scheduler.wait(10);
-      },
-
-      async abort() {
-        throw new Error('boom');
-      },
-    });
-
-    const writer = ws.getWriter();
-    const write = writer.write('test');
-
-    await rejects(writer.abort(), { message: 'boom' });
-
-    await write;
-
-    // The write should reject with undefined because the writer.abort() above
-    // specified undefined. The abort algorithm should not be called again.
-    await rejects(writer.write('should fail'), (err) => err === undefined);
-  },
-};
-
-// Test WritableStream releaseLock while aborting should reject closed promise
-export const writableStreamReleaseLockWhileAborting = {
-  async test() {
-    const ws = new WritableStream({
-      async write() {
-        await scheduler.wait(10);
-      },
-    });
-
-    const writer = ws.getWriter();
-    writer.write('test');
-
-    writer.abort();
-    const closed = writer.closed;
-
-    writer.releaseLock();
-
-    await rejects(closed, TypeError);
-  },
-};
-
-// Test WritableStream throw during in flight close rejects abort and closed promise
-export const writableStreamCloseThrowRejectsPromises = {
-  async test() {
-    const ws = new WritableStream({
-      async close() {
-        throw new Error('boom');
-      },
-    });
-
-    const writer = ws.getWriter();
-    const close = writer.close();
-    const abort = writer.abort();
-    const closed = writer.closed;
-
-    const res = await Promise.allSettled([close, abort, closed]);
-
-    strictEqual(res[0].status, 'rejected');
-    strictEqual(res[1].status, 'rejected');
-    strictEqual(res[2].status, 'rejected');
-
-    // The close and abort promises are rejected with the error thrown,
-    // the closed promise, however, should be rejected with the reason
-    // given in the abort(), which is defaulted to undefined.
-    strictEqual(res[0].reason.message, 'boom');
-    strictEqual(res[1].reason.message, 'boom');
-    strictEqual(res[2].reason, undefined);
-  },
-};
-
-// Test WritableStream sink abort not called while write or close in flight
-export const writableStreamAbortTiming = {
-  async test() {
-    // Abort waits for start
-    {
-      let started = false;
-      const ws = new WritableStream({
-        async start() {
-          await scheduler.wait(10);
-          started = true;
-        },
-        abort() {
-          ok(started, 'The stream should have started first');
-        },
-      });
-
-      await ws.abort();
-    }
-
-    // Abort waits for write
-    {
-      let writeCompleted = false;
-      const ws = new WritableStream({
-        async write() {
-          await scheduler.wait(10);
-          writeCompleted = true;
-        },
-        abort() {
-          ok(writeCompleted, 'The write should have completed');
-        },
-      });
-
-      const writer = ws.getWriter();
-      const write = writer.write('hello');
-      const abort = writer.abort();
-
-      await Promise.allSettled([write, abort]);
-    }
-
-    // Abort waits for close
-    {
-      let closeCompleted = false;
-      const ws = new WritableStream({
-        async close() {
-          await scheduler.wait(10);
-          closeCompleted = true;
-        },
-        abort() {
-          ok(closeCompleted, 'The close should have completed');
-        },
-      });
-
-      const writer = ws.getWriter();
-      const close = writer.close();
-      const abort = writer.abort();
-
-      await Promise.allSettled([close, abort]);
-    }
-  },
-};
-
-// Test WritableStream abort during write should trigger abort algorithm with close pending
-export const writableStreamAbortWriteClosePending = {
-  async test() {
-    let abortCalled = false;
-    const ws = new WritableStream({
-      async write() {
-        await scheduler.wait(10);
-      },
-      abort() {
-        abortCalled = true;
-      },
-    });
-
-    const writer = ws.getWriter();
-    const write = writer.write('hello');
-    const close = writer.close();
-    const abort = writer.abort();
-
-    const res = await Promise.allSettled([write, close, abort]);
-    ok(abortCalled);
-
-    strictEqual(res[0].status, 'fulfilled'); // Write finishes
-    strictEqual(res[1].status, 'rejected'); // Pending close is aborted
-    strictEqual(res[2].status, 'fulfilled'); // Abort finishes
-  },
-};
-
-// Test WritableStream ready promise rejects on controller error not waiting for in flight write
-export const writableStreamErrorDuringInFlightWrite = {
-  async test() {
-    let controller;
-    const ws = new WritableStream({
-      start(c) {
-        controller = c;
-      },
-      async write() {
-        await scheduler.wait(10);
-      },
-    });
-
-    const writer = ws.getWriter();
-
-    const write = writer.write('hello').catch(() => {});
-
-    controller.error('boom');
-
-    await Promise.all([write, writer.ready.catch(() => {})]);
-  },
-};
-
-// Test WritableStream start errors after abort, close rejects
-export const writableStreamStartErrorAfterAbort = {
-  async test() {
-    const ws = new WritableStream({
-      async start() {
-        await scheduler.wait(10);
-        throw new Error('boom');
-      },
-    });
-
-    ws.abort();
-
-    // close() should return a rejected promise. The abort was called
-    // before start finished, so the stream enters an errored state.
-    await ws.close().catch(() => {});
-
-    // Verify the stream is errored
-    strictEqual(ws.locked, false);
-  },
-};
-
-// Test WritableStream rejected sink write does not prevent sink abort
-export const writableStreamRejectedWriteNoPreventAbort = {
-  async test() {
-    let abortCalled = false;
-    const ws = new WritableStream({
-      write() {
-        throw new Error('boom');
-      },
-      abort() {
-        abortCalled = true;
-      },
-    });
-
-    const writer = ws.getWriter();
-    const write = writer.write('hello');
-    const abort = writer.abort();
-
-    const res = await Promise.allSettled([write, abort]);
-    ok(abortCalled);
-    strictEqual(res[0].status, 'rejected');
-    strictEqual(res[1].status, 'fulfilled');
-  },
-};
-
-// Test WritableStream aborted twice
-export const writableStreamAbortedTwice = {
-  async test() {
-    const ws = new WritableStream();
-    const abort1 = ws.abort();
-    const abort2 = ws.abort();
-
-    await Promise.all([abort1, abort2]);
-  },
-};
-
-// Test WritableStream aborting errored stream rejects with stored error
-export const writableStreamAbortOnErroredResolves = {
-  async test() {
-    const ws = new WritableStream({
-      start(c) {
-        c.error(new Error('boom'));
-      },
-    });
-
-    // When aborting an already-errored stream, abort() rejects with the stored error
-    await rejects(ws.abort(), Error);
-  },
-};
-
-// Test WritableStream sink abort not called if stream errored before abort
-export const writableStreamSinkAlgNoCallErrorBeforeAbort = {
-  test() {
-    let controller;
-    let abortCalled = false;
-    const ws = new WritableStream({
-      start(c) {
-        controller = c;
-      },
-      abort() {
-        abortCalled = true;
-      },
-    });
-
-    controller.error(new Error('boom'));
-    ws.abort(new Error('bang'));
-
-    ok(!abortCalled);
-  },
-};
-
-// Test WritableStream writer with pending abort, ready should reject
-export const writableStreamWriterWithPendingAbort = {
-  async test() {
-    const ws = new WritableStream();
-    ws.abort(new Error('boom'));
-    const writer = ws.getWriter();
-
-    await rejects(writer.ready, Error);
-  },
-};
-
-// Test WritableStream promises resolved in order
-export const writableStreamPromisesResolvedInOrder = {
-  async test() {
-    // Write before close
-    {
-      let closeFinished = false;
-      const ws = new WritableStream();
-
-      const writer = ws.getWriter();
-
-      const write = writer.write('hello').then(() => ok(!closeFinished));
-      const close = writer.close();
-      const closed = writer.closed.then(() => (closeFinished = true));
-
-      // Closed promise should not resolve before fulfilled write.
-      await Promise.allSettled([write, close, closed]);
-    }
-
-    // Rejected write before close
-    {
-      let closeFinished = false;
-      let writeFailed = false;
-      const ws = new WritableStream({
-        write() {
-          throw new Error('boom');
-        },
-      });
-
-      const writer = ws.getWriter();
-
-      const write = writer.write('hello').catch(() => {
-        ok(!closeFinished);
-        writeFailed = true;
-      });
-      const close = writer.close();
-      const closed = writer.closed.then(() => (closeFinished = true));
-
-      // Closed promise should not resolve before rejected write.
-      await Promise.allSettled([write, close, closed]);
-      ok(writeFailed);
-    }
-
-    // Writes resolved in order when aborting
-    {
-      const order = [];
-      const ws = new WritableStream({
-        async write() {
-          await scheduler.wait(10);
-        },
-      });
-
-      const writer = ws.getWriter();
-
-      const write1 = writer.write('hello').then(() => order.push(1));
-      const write2 = writer.write('hello').catch(() => order.push(2));
-      const write3 = writer.write('hello').catch(() => order.push(3));
-      const abort = writer.abort();
-
-      await Promise.allSettled([write1, write2, write3, abort]);
-
-      strictEqual(order[0], 1);
-      strictEqual(order[1], 2);
-      strictEqual(order[2], 3);
-    }
-  },
-};
-
-// =====================================================================================
 // Misc tests
 // =====================================================================================
 
@@ -2378,7 +1402,6 @@ export const writableStreamPromisesResolvedInOrder = {
 export const highWaterMarkValidated = {
   test() {
     [-1, -Infinity, NaN, {}, 'foo'].forEach((highWaterMark) => {
-      throws(() => new WritableStream(undefined, { highWaterMark }), TypeError);
       throws(() => new ReadableStream(undefined, { highWaterMark }), TypeError);
     });
   },
@@ -2438,26 +1461,6 @@ export const byobreaderRegression = {
     const reader = rs.getReader({ mode: 'byob' });
     const { done } = await reader.read(new Uint8Array(10));
     ok(done);
-  },
-};
-
-// Test writer double close
-export const writerDoubleClose = {
-  async test() {
-    const ws = new WritableStream({
-      write() {},
-    });
-    const writer = ws.getWriter();
-
-    writer.write(123);
-
-    writer.close();
-    // With capture_async_api_throws, async methods return rejected promises instead of throwing
-    if (Cloudflare.compatibilityFlags.capture_async_api_throws) {
-      await rejects(writer.close(), TypeError);
-    } else {
-      throws(() => writer.close(), TypeError);
-    }
   },
 };
 
@@ -2536,40 +1539,6 @@ export const readableStreamReferencesHold = {
         ok(!done);
         strictEqual(value, 'hello');
       }
-    }
-  },
-};
-
-// Test WritableStream object references are held through gc
-export const writableStreamGc = {
-  async test() {
-    let controller;
-    let writer;
-    let write;
-
-    {
-      const ws = new WritableStream({
-        start(c) {
-          controller = c;
-        },
-      });
-      writer = ws.getWriter();
-    }
-
-    await scheduler.wait(10);
-    gc();
-
-    {
-      write = writer.write(1);
-      writer = undefined;
-    }
-
-    await scheduler.wait(10);
-    gc();
-
-    {
-      await write;
-      strictEqual(controller.signal.aborted, false);
     }
   },
 };
